Update enforcement_readiness.py
Re-write in order to improve scalability. v2 replaces the old logic of getting events from server 1 hostname and interval at a time and instead uses the 'logging/exechistories' endpoint to bulk download all events for all agents for the chosen Policy Group and then do all processing/counting locally. In testing this new version was able to download and process 6.1M events for 95K agents with a 31 minute runtime.